* niden-code-v6.0.x: (36 commits)
  [#6.0.x] - correcting tests
  [#6.0.x] - correcting Request
  [#6.0.x] - fixing deprecation warnings
  [#6.0.x] - upgrades to actionsg
  [#6.0.x] - path correction
  [#6.0.x] - additional corrections for tests
  [#6.0.x] - stan corrections
  [#6.0.x] - more phpcs
  [#6.0.x] - phpcs
  [#6.0.x] - phpcs
  [#6.0.x] - additional adjustments
  [#6.0.x] - adjusting code with cphalcon
  [#6.0.x] - database tests alignment with cphalcon
  [#6.0.x] - datamapper tests alignment
  [#6.0.x] - unit test alignment from cphalcon
  [#6.0.x] - implementing deleteMultiple for storage and cache
  [#6.0.x] - aligning appliation with cphalcon
  [#6.0.x] - aligning assets with cphalcon
  [#6.0.x] - alignment with cphalcon
  [#6.0.x] - aligning autoload with cphalcon
  ...

+++ src/Config/Adapter/Ini.php
@@ -157,7 +157,7 @@ class Ini extends Config
         // Decode float/int
         if (is_numeric($ini)) {
             if (preg_match('/[.]+/', $ini)) {
-                return (double)$ini;
+                return (float)$ini;
             }
 
             return (int)$ini;
