+++ src/Filter/Sanitize/FloatVal.php
@@ -30,7 +30,7 @@ class FloatVal
      */
     public function __invoke(mixed $input): float
     {
-        return (double)filter_var(
+        return (float)filter_var(
             $input,
             FILTER_SANITIZE_NUMBER_FLOAT,
             [
